modified: separate models from code
